g++12.2: cleanup compiler warnings
Here is a couple of small cleanups to reduce the noise from:
    g++ -W -Wall cppfront.cpp -std=c++20 -o cppfront

To aid seeing the wood for the trees, the actual invocation was:
   g++ -W -Wall -Wno-unused-parameter -Wno-unused-variable ....

Selectively drop those "-Wno-" options to get lots of noise.

That leaves 5 warnings for your consideration:

1)
   parse.h:1416:38: warning: missing initializer for member
       ‘cpp2::capture::str’ [-Wmissing-field-initializers]
   1416 |                 n->cap_grp->push_back({n.get()});

2)
  parse.h:1419:62: warning: missing initializer for member
       ‘cpp2::postfix_expression_node::term::id_expr’
       [-Wmissing-field-initializers]
  1419 |             auto term = postfix_expression_node::term{&curr()};

3)
  parse.h:1419:62: warning: missing initializer for member
       ‘cpp2::postfix_expression_node::term::op_close’
       [-Wmissing-field-initializers]

4)
  parse.h: In member function ‘std::unique_ptr<cpp2::declaration_node>
       cpp2::parser::unnamed_declaration(cpp2::source_position, bool, bool)’:

5)
  parse.h:2803:14: warning: variable ‘start_pos’ set but not used
       [-Wunused-but-set-variable]
  2803 |         auto start_pos = pos;

This was run on Ubuntu 22.10 .

+++ source/common.h
@@ -396,7 +396,7 @@ public:
             }
             print("  -");
             auto n = flag.name.substr(0, flag.unique_prefix);
-            if (flag.unique_prefix < flag.name.length()) {
+            if (flag.unique_prefix < static_cast<ssize_t>(flag.name.length())) {
                 n += "[";
                 n += flag.name.substr(flag.unique_prefix);
                 n += "]";
